Create a renderer class which will handle rendering out the stats.
This completely disconnects the stats rendering from the image rendering
process. At the class level it keeps track of the registered hooks. It now has
the register_hooks function as well as the newly added get_hooks function
which lists all the registered hooks(used in the layout).

At the instance level it keeps track of the active hooks. It also has the
rendering function which handles actually rendering the stats file.