Introduce hook groups.
Hook groups is a way of grouping together related hooks in a meaningful
way.  Currently they are not a separate class, but instead simply a
dictionary mapping `idname` => ('label', 'priority') idname and label are
both strings, priority is an integer.  The priority is automatically generated
based on the order of registration. That may change in the future.

The registration command looks like this:
SRDRenderer.register_group('idname', 'label')